requirements.txt
pytest
pytest-icdiff
pytest-cov
pylint
pandas
streamlit
python-dotenv
psycopg[binary]
psycopg2-binary
altair[all]
plotly
streamlit-pdf-viewer
boto3
flask
aiohttp
opencage
nbformat
dotenv
awswrangler
xhtml2pdf
numpy
